requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
psycopg2-binary==2.9.7
tensorflow==2.13.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
Werkzeug==2.3.7
python-dotenv==1.0.0
itsdangerous==2.2.0
h5py==3.9.0